+++ brutils/__init__.py
@@ -47,11 +47,12 @@ from brutils.date import convert_date_to_text
 
 # Email Import
 from brutils.email import is_valid as is_valid_email
-from brutils.ibge.municipality import (
-    get_municipality_by_code,
-)
 
 # IBGE Imports
+from brutils.ibge.municipality import (
+    get_code_by_municipality_name,
+    get_municipality_by_code,
+)
 from brutils.ibge.uf import (
     convert_code_to_uf,
 )
@@ -181,4 +182,5 @@ __all__ = [
     # IBGE
     "convert_code_to_uf",
     "get_municipality_by_code",
+    "get_code_by_municipality_name",
 ]